feat: Simplifies submodule dependencies by inlining and cleanup
- Inlines lib/ submodules
- Removes unused circuits/ submodules

Decoupling logic via submodules decouples purely on a technical level
but strongly binds it logically. So, by reusing the web-server library
as a submodule, one tightly couples together all the users of this
web-server. This means that, whenever you modify this library, it
triggers cascading changes to all the dependencies. This might be
expected for a web-server but definately not for things like nvs_init
and wifiinit. These are device specific configurations not to be shared
across multiple devices.