feat: adds pipeline embedded build step
**Whats the current behavior**

Since development is done almost exclusively in the `native-dev` environment, breakages in the embedded library structure can easily be missed. The CI pipeline builds the application for `native-dev` only.

**Whats the new behavior**

The `build_env.ps1` has been improved to accespt multiple environments so that the `gitlab_ci.yml` is able to build the application for both embedded and native.

closes [#149]

[#149]: https://gitlab.com/ohsnaparts/osasoftworks/zube/presence-cube/-/issues/149